### 2025-10-07 #9169 @editorialbot
[PRE REVIEW]: ThRasE: A QGIS Plugin for Manual Post-Classification Correction and Quality Assurance of Thematic Maps
**Submitting author:** @XavierCLL (<a href="http://orcid.org/0000-0001-7593-6477">Xavier Corredor Llano</a>)
**Repository:** https://github.com/SMByC/ThRasE
**Branch with paper.md** (empty if default branch): main
**Version:** 25.8
