### 2025-10-07 2fe269e
Improving README file and typos
