### 2025-10-08 15134ca
Update DOI in bibliography entries
