+++ paper/paper.bib
@@ -8,7 +8,7 @@
     pages = {387-XXXIII},
     year = {2007},
     isbn = {978-0-12-369407-2},
-    doi = {https://doi.org/10.1016/B978-012369407-2/50012-7},
+    doi = {10.1016/B978-012369407-2/50012-7},
     url = {https://www.sciencedirect.com/science/article/pii/B9780123694072500127},
     author = {Robert A. Schowengerdt},
 }
@@ -18,7 +18,8 @@
   journal={Cartographic perspectives},
   number={13},
   pages={10--19},
-  year={1992}
+  year={1992},
+  doi={10.14714/CP13.1000}
 }
 @article{manandhar:2009,
   title={Improving the accuracy of land use and land cover classification of Landsat data using post-classification enhancement},
@@ -28,7 +29,8 @@
   number={3},
   pages={330--344},
   year={2009},
-  publisher={Molecular Diversity Preservation International}
+  publisher={Molecular Diversity Preservation International},
+  doi={10.3390/rs1030330}
 }
 @article{thakkar:2017,
   title={Post-classification corrections in improving the classification of Land Use/Land Cover of arid region using RS and GIS: The case of Arjuni watershed, Gujarat, India},
@@ -38,7 +40,8 @@
   number={1},
   pages={79--89},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ejrs.2016.11.006}
 }
 @article{hutchinson:1982,
   title={Classification improvement},
@@ -57,7 +60,8 @@
   number={3},
   pages={389--398},
   year={2001},
-  publisher={Sage Publications Sage CA: Thousand Oaks, CA}
+  publisher={Sage Publications Sage CA: Thousand Oaks, CA},
+  doi={10.1177/030913330102500306}
 }
 @article{lightfoot:1987,
   title={Recognition and assessment of error in geographic information systems},
@@ -72,13 +76,15 @@
   title={Measurements from maps: principles and methods of cartometry},
   author={Maling, Derek H},
   year={2013},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/0160-9327(89)90096-3}
 }
 @book{congalton:2019,
   title={Assessing the accuracy of remotely sensed data: principles and practices},
   author={Congalton, Russell G and Green, Kass},
   year={2019},
-  publisher={CRC press}
+  publisher={CRC press},
+  doi={10.1201/9781420048568}
 }
 @article{vallet:2024,
   title={Generating high-resolution land use and land cover maps for the greater Mari{\~n}o watershed in 2019 with machine learning},
@@ -88,7 +94,8 @@
   number={1},
   pages={915},
   year={2024},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41597-024-03750-x}
 }
 @article{biljecki:2021,
   title={Street view imagery in urban analytics and GIS: A review},
@@ -97,7 +104,8 @@
   volume={215},
   pages={104217},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.landurbplan.2021.104217}
 }
 @article{sishodia:2020,
   title={Applications of remote sensing in precision agriculture: A review},
@@ -107,7 +115,8 @@
   number={19},
   pages={3136},
   year={2020},
-  publisher={MDPI}
+  publisher={MDPI},
+  doi={10.3390/rs12193136}
 }
 @article{bhatt:2012,
   title={GIS in epidemiology: applications and services},
[…]
