### 2025-10-08 269aec7
Adjust and reducing paper content
