+++ paper/paper.md
@@ -21,15 +21,9 @@ date: 05 September 2025
 bibliography: paper.bib
 ---
 
-# Summary
-
-Thematic maps are critical tools that translate complex spatial data into actionable insights, but their reliability depends on rigorous quality assurance because mapping workflows inevitably introduce errors and uncertainty. To address this need, we developed ThRasE, an open-source QGIS plugin that provides comprehensive editing tools, modification tracking, systematic visual inspection features, and capabilities for comparison with reference data for manual post-classification correction of thematic maps.
-
 # Statement of need
 
-Thematic maps serve as indispensable analytical tools that simplify complex, continuous spatial data into visually accessible representations, enabling informed decision-making across a wide range of disciplines. In environmental management, land-use/land-cover maps underpin risk assessment, resource stewardship, and scenario planning [@vallet:2024]. In cities, thematic mapping enables infrastructure and land-use analysis, green-space equity assessments, and data-driven urban analytics [@biljecki:2021]. In agriculture, map-based precision farming uses yield, vegetation, and moisture layers to guide variable-rate inputs and crop monitoring [@sishodia:2020]. In public health, they highlight disease or exposure hotspots for precise responses, enabling hotspot detection in epidemiology [@bhatt:2012]. Overall, thematic maps function as instruments for representation, communication, and decision support, providing a symbolic representation of reality that bridges complex data with practical insights.
-
-Ensuring the quality of thematic maps enhances the reliability of spatial data, which is critical for supporting informed decision-making [@foody:2001]. Because maps are models of reality, they inevitably contain errors introduced by projection, compilation, and symbolization [@lightfoot:1987; @maling:2013]. Many issues originate in the production process itself. For example, satellite-based classifications (e.g., supervised or unsupervised workflows) can introduce errors of omission or commission, scale mismatches, or temporal misalignments with ground conditions [@Schowengerdt:2007]. These limitations do not undermine the value of thematic maps; rather, they highlight the importance of rigorous quality assurance. Best practices include conducting formal accuracy assessments [@congalton:2019], transparently reporting error and uncertainty [@maceachren:1992], and, where appropriate, applying targeted post-classification corrections to address misclassifications [@manandhar:2009].
+Thematic maps serve as indispensable analytical tools that simplify complex, continuous spatial data into visually accessible representations for a wide range of disciplines and applications. However, mapping workflows inevitably introduce errors, making quality assurance essential for reliable spatial data and informed decision-making [@foody:2001]. Because maps are models of reality, they inevitably contain errors introduced by projection, compilation, and symbolization [@lightfoot:1987; @maling:2013]. Many issues originate in the production process itself. For example, satellite-based classifications (e.g., supervised or unsupervised workflows) can introduce errors of omission or commission, scale mismatches, or temporal misalignments with ground conditions [@Schowengerdt:2007]. These limitations do not undermine the value of thematic maps; rather, they highlight the importance of rigorous quality assurance. Best practices include conducting formal accuracy assessments [@congalton:2019], transparently reporting error and uncertainty [@maceachren:1992], and, where appropriate, applying targeted post-classification corrections to address misclassifications [@manandhar:2009].
 
 The value of a thematic map ultimately depends on its quality, which is primarily driven by the original data sources and classification methodologies. However, post-classification correction serves as a widely adopted final step to refine misclassifications and improve outcomes [@hutchinson:1982]. A common approach, particularly for remote sensing-based maps, involves integrating ancillary data and knowledge-based rules to resolve misclassifications, reduce commission errors, and enhance overall accuracy [@manandhar:2009; @thakkar:2017]. These corrections often rely on manual editing or semi-automated techniques guided by expert knowledge and reference data. This is precisely the context in which ThRasE was developed.
 
