### 2025-10-08 #9169 @XavierCLL
Hi @kthyng thanks so much for the review!

I am going to respond to you item by item:

- [x] **DOI** fixed (check previous bot result)
- [x] **Installation instructions** here https://github.com/SMByC/ThRasE?tab=readme-ov-file#install and demo files https://smbyc.github.io/ThRasE/example.html
- [x] **Documentation exists at all** https://smbyc.github.io/ThRasE/
- [x] **Tests of some sort** critical test of the plugin are here https://github.com/SMByC/ThRasE/blob/main/.github/workflows/tests.yml and https://github.com/SMByC/ThRasE/blob/main/tests/test_editing.py
- [ ] check for length [250-1000 words] I removed the summary and combined two paragraphs in one to reduce it as much as I can, last counting was 1113, is that ok?
- [x] **Substantial scholarly effort** the plugin has been developing for more than 6 years, used for different research and projects
- [x] **Clear research** application check the last section of "Statement of need" in the paper

Thanks,
Xavier
