+++ pyproject.toml
@@ -1,6 +1,44 @@
+[project]
+name = "thrase"
+version = "25.8"
+description = "Open-source QGIS plugin for rigorous editing and quality assurance of thematic rasters."
+readme = "README.md"
+requires-python = ">=3.9"
+license = "GPL-3.0-only"
+authors = [{ name = "Xavier C. Llano", email = "xavier.corredor.llano@gmail.com" }]
+keywords = ["qgis", "raster", "thematic", "categorical", "editing", "reclassification", "navigation", "quality"]
+classifiers = [
+  "Development Status :: 5 - Production/Stable",
+  "Programming Language :: Python",
+  "Topic :: Scientific/Engineering :: GIS",
+]
+dependencies = [
+  "numpy",
+  "PyYAML",
+]
+
+[project.optional-dependencies]
+# uv pip sync pyproject.toml --extra docs
+docs = [
+  "jupyter-book",
+  "jupytext",
+]
+# uv pip sync pyproject.toml --extra test
+test = [
+  "pytest",
+  "pytest-qgis",
+]
+
+[project.urls]
+Homepage = "https://smbyc.github.io/ThRasE"
+Documentation = "https://smbyc.github.io/ThRasE"
+Repository = "https://github.com/SMByC/ThRasE"
+Issues = "https://github.com/SMByC/ThRasE/issues"
+"QGIS Plugin" = "https://plugins.qgis.org/plugins/ThRasE/"
+
 [tool.pyright]
 extraPaths = [".."]  # include project folder
-# venv = ".venv"
+venv = ".venv"
 reportUnknownVariableType = "none"
 reportUnknownParameterType = "none"
 reportUnknownMemberType = "none"
