### 2025-10-20 ba11bc4
Refactor YAML loading and dumping to support legacy formats and improve the default format for storing configuration
