### 2025-10-20 92dc18b
Preserve all metadata of the original file when is using the global editing tool function to apply to the whole image with gdal
