+++ core/editing.py
@@ -367,9 +367,12 @@ class LayerToEdit(object):
         """Edit the whole image with the new values using gdal"""
         try:
             # read
-            ds_in = gdal.Open(self.file_path)
+            ds_in = gdal.Open(self.file_path, gdal.GA_ReadOnly)
+            if ds_in is None:
+                raise RuntimeError(f"Unable to open raster {self.file_path}")
             num_bands = ds_in.RasterCount
-            data_array = ds_in.GetRasterBand(self.band).ReadAsArray().astype(int)
+            src_band = ds_in.GetRasterBand(self.band)
+            data_array = src_band.ReadAsArray().astype(int)
             new_data_array = deepcopy(data_array)
 
             # apply changes
@@ -377,27 +380,116 @@ class LayerToEdit(object):
                 new_data_array[data_array == old_value] = new_value
             del data_array
 
+            def safe_call(method, *args):
+                try:
+                    method(*args)
+                except Exception:
+                    pass
+
+            def copy_band_metadata(src, dst):
+                nodata = src.GetNoDataValue()
+                if nodata is not None:
+                    safe_call(dst.SetNoDataValue, nodata)
+
+                color_table = src.GetRasterColorTable()
+                if color_table is not None:
+                    safe_call(dst.SetRasterColorTable, color_table.Clone())
+
+                category_names = src.GetCategoryNames()
+                if category_names:
+                    safe_call(dst.SetCategoryNames, category_names)
+
+                rat = src.GetDefaultRAT()
+                if rat is not None:
+                    safe_call(dst.SetDefaultRAT, rat.Clone())
+
+                metadata_domains = src.GetMetadataDomainList()
+                if not metadata_domains:
+                    metadata = src.GetMetadata()
+                    if metadata:
+                        safe_call(dst.SetMetadata, metadata)
+                else:
+                    for domain in metadata_domains:
+                        metadata = src.GetMetadata(domain)
+                        if metadata:
+                            safe_call(dst.SetMetadata, metadata, domain)
+
+                unit = src.GetUnitType()
+                if unit:
+                    safe_call(dst.SetUnitType, unit)
+
+                scale = src.GetScale()
+                if scale is not None:
+                    safe_call(dst.SetScale, scale)
+
+                offset = src.GetOffset()
+                if offset is not None:
+                    safe_call(dst.SetOffset, offset)
+
+                description = src.GetDescription()
+                if description:
+                    safe_call(dst.SetDescription, description)
+
+                safe_call(dst.SetColorInterpretation, src.GetColorInterpretation())
+
+            def copy_dataset_metadata(src, dst):
+                metadata_domains = src.GetMetadataDomainList()
+                if not metadata_domains:
+                    metadata = src.GetMetadata()
+                    if metadata:
+                        safe_call(dst.SetMetadata, metadata)
+                else:
+                    for domain in metadata_domains:
+                        metadata = src.GetMetadata(domain)
+                        if metadata:
+                            safe_call(dst.SetMetadata, metadata, domain)
+
+                description = src.GetDescription()
+                if description:
+                    safe_call(dst.SetDescription, description)
+
+                gcps = src.GetGCPs()
+                if gcps:
+                    safe_call(dst.SetGCPs, gcps, src.GetGCPProjection())
+
             # create file
             fn, ext = os.path.splitext(self.file_path)
             fn_out = fn + "_tmp" + ext
-            driver = gdal.GetDriverByName(ds_in.GetDriver().ShortName)
+            driver_name = ds_in.GetDriver().ShortName
[…]
