### 2025-10-21 a98a578
Add the option to record the changes in ThRasE registry in dialog for appling the changes to the entire image tool
