### 2025-10-22 #9169 @XavierCLL
Dear @kthyng

Yes, sorry. Someone familiar with QGIS knows how to install a plugin, but it would be better to provide clearer instructions on how to do it. Fixed: https://github.com/SMByC/ThRasE?tab=readme-ov-file#install

Please let me know what else you need to continue with the pre-review process

Thanks you!
