### 2025-10-22 a736b56
Enhance some functionality in ApplyFromThematicClasses dialog with registry status tooltip and improved pixel processing
