### 2025-10-23 92ef3ab
Moving Gdal copy metadata common functions to others_utils
