### 2025-10-25 310c112
Refactoring the global editing tool applying changes using selected classes from another file using gdal improving performance and functionality
