+++ gui/apply_from_thematic_classes.py
@@ -23,18 +23,19 @@ import os
 import uuid
 from copy import deepcopy
 from pathlib import Path
+from shutil import move
 import numpy as np
 from osgeo import gdal
 
-from qgis.core import QgsMapLayerProxyModel, Qgis, QgsPointXY
+from qgis.core import QgsMapLayerProxyModel, Qgis
 from qgis.gui import QgsMapToolPan
 from qgis.PyQt import uic
 from qgis.PyQt.QtGui import QColor
 from qgis.PyQt.QtWidgets import QDialog, QFileDialog, QTableWidgetItem, QDialogButtonBox
 from qgis.PyQt.QtCore import pyqtSlot, Qt
 
-from ThRasE.core.editing import Pixel, LayerToEdit, edit_layer
-from ThRasE.utils.others_utils import get_xml_style
+from ThRasE.core.editing import Pixel, PixelLog, LayerToEdit
+from ThRasE.utils.others_utils import get_xml_style, copy_band_metadata, copy_dataset_metadata
 from ThRasE.utils.qgis_utils import load_and_select_filepath_in, apply_symbology, get_file_path_of_layer
 from ThRasE.utils.system_utils import block_signals_to, error_handler, wait_process
 
@@ -183,10 +184,10 @@ class ApplyFromThematicClasses(QDialog, FORM_CLASS):
                 value = xml_item.get("value")
                 color = xml_item.get("color")
                 alpha = xml_item.get("alpha")
-                
+
                 if value is None or color is None or alpha is None:
                     continue  # Skip items with missing attributes
-                
+
                 pixel = {"value": int(value), "color": {}, "select": False}
                 item_color = color.lstrip('#')
                 item_color = tuple(int(item_color[i:i + 2], 16) for i in (0, 2, 4))
@@ -198,7 +199,7 @@ class ApplyFromThematicClasses(QDialog, FORM_CLASS):
             except (ValueError, AttributeError, TypeError):
                 # Skip items with invalid data
                 continue
-        
+
         self.pixel_classes_backup = deepcopy(self.pixel_classes)
 
         self.set_pixel_table()
@@ -276,8 +277,11 @@ class ApplyFromThematicClasses(QDialog, FORM_CLASS):
             apply_symbology(self.thematic_file_classes, band, symbology)
 
     @wait_process
-    @edit_layer
     def apply(self):
+        """Apply changes to the entire thematic map using a different thematic classes map"""
+
+        from ThRasE.thrase import ThRasE
+
         pixel_table = self.PixelTable
         if pixel_table.rowCount() == 0:
             self.MsgBar.pushMessage("Error: pixel table for class selection is empty", level=Qgis.Warning, duration=10)
@@ -290,47 +294,190 @@ class ApplyFromThematicClasses(QDialog, FORM_CLASS):
             return
 
         extent_intercepted = LayerToEdit.current.qgs_layer.extent().intersect(self.thematic_file_classes.extent())
-
-        ps_x = self.thematic_file_classes.rasterUnitsPerPixelX()  # pixel size in x
-        ps_y = self.thematic_file_classes.rasterUnitsPerPixelY()  # pixel size in y
-
-        # locate the pixel centroid in x and y for the pixel in min/max in the extent
-        y_min = extent_intercepted.yMinimum() + (ps_y / 2.0)
-        y_max = extent_intercepted.yMaximum() - (ps_y / 2.0)
-        x_min = extent_intercepted.xMinimum() + (ps_x / 2.0)
-        x_max = extent_intercepted.xMaximum() - (ps_x / 2.0)
-        # index for extent intercepted start in the corner left-top respect to thematic file classes
[…]
