### 2025-10-25 d7790bb
Enhance registry checkbox and tooltip in edit the whole image dialog
