### 2025-10-25 1e9f383
Rename from "edit whole image" to "edit entire thematic raster"
