### 2025-10-25 4dd8da5
Update/renaming "Apply Within Selected Classes" in global editing
