### 2025-10-27 1a7e416
Refactor registry management and performance using a memory layer functionality for tile geometries, improving group display handling. Others adjusted behavior in the registry.
