+++ core/registry.py
@@ -32,15 +32,18 @@ from qgis.core import (
     QgsFeature,
     QgsVectorFileWriter,
     QgsCoordinateTransformContext,
+    QgsVectorLayer,
+    QgsSingleSymbolRenderer,
+    QgsFillSymbol,
 )
-from qgis.gui import QgsRubberBand
 
 from ThRasE.utils.system_utils import wait_process
 
 
 class RegistryTile:
-    def __init__(self, idx, center_x, center_y, px_size_x, px_size_y):
+    def __init__(self, idx, group_idx, center_x, center_y, px_size_x, px_size_y, memory_layer):
         self.idx = idx
+        self.group_idx = group_idx
         self.center_x = center_x
         self.center_y = center_y
         self.px_size_x = px_size_x
@@ -53,20 +56,26 @@ class RegistryTile:
 
         self.extent = QgsRectangle(xmin, ymin, xmax, ymax)
 
-        # use fast rectangle geometry; polygon corners will be implicit
-        self.geometry = QgsGeometry.fromRect(self.extent)
+        # create feature in memory layer
+        geom = QgsGeometry.fromRect(self.extent)
+        feature = QgsFeature(memory_layer.fields())
+        feature.setGeometry(geom)
+        feature.setAttributes([self.idx, self.group_idx, center_x, center_y])
+        
+        # add feature to memory layer
+        memory_layer.dataProvider().addFeature(feature)
+        self.feature_id = feature.id()
 
 
 class RegistryTileGroup:
-    def __init__(self, idx, group_id, edit_date, tiles, tile_color):
+    def __init__(self, idx, group_id, edit_date, tiles, memory_layer, registry):
         self.idx = idx
         self.group_id = group_id
         self.edit_date = edit_date
         self.tiles = tiles
-        self.tile_color = tile_color
+        self.memory_layer = memory_layer
+        self.registry = registry
         self.extent = None
-        self.union_geom = None
-        self.rubber_bands = []  # rubber bands for group outline
 
     def tiles_extent(self):
         if self.extent is None:
@@ -76,57 +85,22 @@ class RegistryTileGroup:
             self.extent = extent
         return self.extent
 
-    @wait_process
     def show(self):
-        # draw only the border outline of the group as a single polygon
-        from ThRasE.gui.main_dialog import ThRasEDialog
+        # display tile borders without fill by applying filter to memory layer
         if not self.tiles:
             return
         
-        # compute union geometry once and cache it
-        if self.union_geom is None:
-            if len(self.tiles) == 1:
-                self.union_geom = self.tiles[0].geometry
-            else:
-                geoms = [tile.geometry for tile in self.tiles]
-                self.union_geom = QgsGeometry.unaryUnion(geoms)
-        union_geom = self.union_geom
-        if union_geom is None or union_geom.isEmpty():
-            return
-        
-        # avoid duplicate rubber bands if show() is called again
-        if self.rubber_bands:
-            self.clear()
-
-        # draw the outline in all active view widgets
-        for view_widget in ThRasEDialog.view_widgets:
-            if not view_widget.is_active:
-                continue
-            # create rubber band for the group outline
-            rb = QgsRubberBand(view_widget.render_widget.canvas, QgsWkbTypes.PolygonGeometry)
-            rb.setToGeometry(union_geom, None)
-            rb.setStrokeColor(self.tile_color)
-            rb.setFillColor(QColor(0, 0, 0, 0))  # transparent fill
-            rb.setWidth(3)
-            rb.show()
-            self.rubber_bands.append(rb)
-            view_widget.render_widget.canvas.clearCache()
-            view_widget.render_widget.canvas.refresh()
[…]
