### 2025-10-27 06810d0
Add group_id in the exported registry vector data
