+++ core/registry.py
@@ -355,55 +355,61 @@ class Registry:
             self.current_group.center()
         self.current_group.show()
 
-    def export_pixel_logs(self, output_file_path):
-        """Export all registered pixel edits to a vector file.
+    def export_registry(self, output_file_path):
+        """Export all registered pixel edits (registry) to a vector file.
 
         Each feature is a square polygon representing the pixel modified and
-        includes attributes: old_value, new_value, edit_date (ISO string).
+        includes attributes: group_id, old_value, new_value, edit_date.
 
         Returns (success: bool, message: str, written_count: int).
         """
-        pixel_logs = list(self.layer_to_edit.pixel_log_store.values())
+        if not self.layer_to_edit.pixel_log_store:
+            return False, "No features to export", 0
 
         # ensure output extension and driver
         path, ext = os.path.splitext(output_file_path)
         ext = ext.lower()
         if ext not in [".gpkg", ".shp", ".geojson"]:
-            # default to GeoPackage if no/unknown extension
             output_file_path = path + ".gpkg"
             ext = ".gpkg"
-        driver_by_ext = {
-            ".gpkg": "GPKG",
-            ".shp": "ESRI Shapefile",
-            ".geojson": "GeoJSON",
-        }
-        driver_name = driver_by_ext.get(ext, "GPKG")
+        
+        driver_name = {".gpkg": "GPKG", ".shp": "ESRI Shapefile", ".geojson": "GeoJSON"}.get(ext, "GPKG")
 
-        # fields
+        # create mapping from UUID group_id
+        group_ids = {group.group_id: group.idx for group in self.groups} if self.groups else {}
+
+        # define fields
         fields = QgsFields()
+        fields.append(QgsField("group_id", QVariant.Int))
         fields.append(QgsField("old_value", QVariant.Int))
         fields.append(QgsField("new_value", QVariant.Int))
         fields.append(QgsField("edit_date", QVariant.String))
 
-        # pixel footprint size
+        # pre-calculate pixel size (constant for all features)
         psx = self.layer_to_edit.qgs_layer.rasterUnitsPerPixelX()
         psy = self.layer_to_edit.qgs_layer.rasterUnitsPerPixelY()
+        half_psx = psx / 2
+        half_psy = psy / 2
 
-        # build features
+        # build features efficiently
         features = []
-        for pl in pixel_logs:
+        for pl in self.layer_to_edit.pixel_log_store.values():
             cx = pl.pixel.x()
             cy = pl.pixel.y()
-            xmin = cx - psx / 2
-            xmax = cx + psx / 2
-            ymin = cy - psy / 2
-            ymax = cy + psy / 2
-            rect = QgsRectangle(xmin, ymin, xmax, ymax)
+            
+            # create geometry
+            rect = QgsRectangle(cx - half_psx, cy - half_psy, cx + half_psx, cy + half_psy)
             geom = QgsGeometry.fromRect(rect)
 
+            # create feature
             feat = QgsFeature(fields)
             feat.setGeometry(geom)
-            feat.setAttributes([int(pl.old_value), int(pl.new_value), pl.edit_date.isoformat()])
+            feat.setAttributes([
+                group_ids.get(pl.group_id),
+                int(pl.old_value),
+                int(pl.new_value),
+                pl.edit_date.isoformat()
+            ])
             features.append(feat)
 
         if not features:
@@ -416,14 +422,17 @@ class Registry:
         options.fileEncoding = "UTF-8"
         options.layerName = os.path.splitext(os.path.basename(output_file_path))[0]
         options.actionOnExistingFile = QgsVectorFileWriter.CreateOrOverwriteFile
-        transform_context = QgsCoordinateTransformContext()
 
[…]
