### 2025-10-27 57fdf39
Fix registry layer redering when others layers changes in the render widget
