### 2025-10-28 1a0cc15
Fix undo/redo actions for the registry memory layer. Overall, improve performance when adding and removing groups in the registry during large editing sessions
