+++ core/registry.py
@@ -2,7 +2,7 @@
 """
 /***************************************************************************
  ThRasE
- 
+
  A powerful and fast thematic raster editor Qgis plugin
                               -------------------
         copyright            : (C) 2019-2025 by Xavier Corredor Llano, SMByC
@@ -18,7 +18,7 @@
  *                                                                         *
  ***************************************************************************/
 """
-
+import itertools
 import os
 
 from qgis.PyQt.QtGui import QColor
@@ -61,7 +61,7 @@ class RegistryTile:
         feature = QgsFeature(memory_layer.fields())
         feature.setGeometry(geom)
         feature.setAttributes([self.idx, self.group_idx, center_x, center_y])
-        
+
         # add feature to memory layer
         memory_layer.dataProvider().addFeature(feature)
         self.feature_id = feature.id()
@@ -89,7 +89,7 @@ class RegistryTileGroup:
         # display tile borders without fill by applying filter to memory layer
         if not self.tiles:
             return
-        
+
         # set filter to show only tiles from this group
         filter_expr = f'"group_idx" = {self.idx}'
         self.memory_layer.setSubsetString(filter_expr)
@@ -108,11 +108,11 @@ class RegistryTileGroup:
         tiles_extent = self.tiles_extent()
         if tiles_extent.isEmpty():
             return
-        
+
         # get the center point of the group
         center_x = tiles_extent.center().x()
         center_y = tiles_extent.center().y()
-        
+
         # center all active views on this point without changing scale
         for view_widget in ThRasEDialog.view_widgets:
             if not view_widget.is_active:
@@ -135,11 +135,11 @@ class Registry:
         self.current_group = None
         self.tiles_color = QColor("#ff00ff")
         self.enabled = True
-        
+
         # create memory vector layer for tiles
         self.memory_layer = None
         self.create_memory_layer()
-        
+
         # single renderer for both display modes
         self.renderer = None
         self.setup_renderer()
@@ -148,11 +148,11 @@ class Registry:
         """Create memory vector layer to store tile geometries."""
         crs = self.layer_to_edit.qgs_layer.crs()
         self.memory_layer = QgsVectorLayer(
-            f"Polygon?crs={crs.authid()}", 
-            "ThRasE Registry", 
+            f"Polygon?crs={crs.authid()}",
+            "ThRasE Registry",
             "memory"
         )
-        
+
         # add fields
         provider = self.memory_layer.dataProvider()
         provider.addAttributes([
@@ -162,19 +162,19 @@ class Registry:
             QgsField("center_y", QVariant.Double),
         ])
         self.memory_layer.updateFields()
-        
+
         # set initial state (hidden by default)
         self.memory_layer.setSubsetString("FALSE")  # hide all initially
 
     def setup_renderer(self, color=None):
         """Setup single renderer for draw the registry tiles.
-        
+
         Args:
             color: QColor to use for border. If None, uses the default color
         """
         if color is None:
             color = self.tiles_color
-        
+
         # Simple border renderer - no fill, 0.3 border width
         border_symbol = QgsFillSymbol.createSimple({
             'color': 'transparent',
@@ -184,7 +184,7 @@ class Registry:
             'outline_style': 'solid'
         })
         self.renderer = QgsSingleSymbolRenderer(border_symbol)
-    
+
     def update_registry_layer_in_canvases(self):
         """Refresh render layers in all active canvases to update registry layer visibility."""
         from ThRasE.gui.main_dialog import ThRasEDialog
@@ -196,7 +196,7 @@ class Registry:
         self.clear()
         self.groups = []
         self.current_group = None
-        
+
         # clear memory layer
[…]
