+++ gui/render_widget.py
@@ -62,45 +62,48 @@ class RenderWidget(QWidget):
 
     def update_render_layers(self):
         with block_signals_to(self):
-            from ThRasE.core.editing import LayerToEdit
             # set the CRS of the canvas view
             if self.crs:
-                # use the crs of thematic layer to edit
                 self.canvas.setDestinationCrs(self.crs)
             else:
-                # use the crs set in Qgis
                 self.canvas.setDestinationCrs(iface.mapCanvas().mapSettings().destinationCrs())
+            
             # get all valid activated layers
             valid_layers = [layer_toolbar.layer for layer_toolbar in self.layer_toolbars
                            if layer_toolbar.is_active and layer_toolbar.layer is not None
                            and layer_toolbar.layer.isValid()]
-            if len(valid_layers) == 0:
+            
+            if not valid_layers:
                 self.canvas.setLayers([])
                 self.refresh()
                 return
-            # include registry memory layer if active
-            if LayerToEdit.current and LayerToEdit.current.registry.memory_layer:
-                memory_layer = LayerToEdit.current.registry.memory_layer
-                # add registry layer on top of other layers
-                self.canvas.setLayers([memory_layer] + valid_layers)
-            else:
-                # set to canvas without registry layer
-                self.canvas.setLayers(valid_layers)
+            
+            # include registry memory layer if exists
+            from ThRasE.core.editing import LayerToEdit
+            memory_layer = (LayerToEdit.current.registry.memory_layer 
+                           if LayerToEdit.current and LayerToEdit.current.registry.memory_layer 
+                           else None)
+            
+            layers_to_set = [memory_layer] + valid_layers if memory_layer else valid_layers
+            self.canvas.setLayers(layers_to_set)
+            
             # set init extent from other view if any is activated else set layer extent
             from ThRasE.gui.main_dialog import ThRasEDialog
-            others_extents = [view_widget.render_widget.canvas.extent() for view_widget in ThRasEDialog.view_widgets
-                              if view_widget.is_active and view_widget.render_widget != self
-                              and not view_widget.render_widget.canvas.extent().isEmpty()]
+            others_extents = [view_widget.render_widget.canvas.extent() 
+                             for view_widget in ThRasEDialog.view_widgets
+                             if view_widget.is_active and view_widget.render_widget != self
+                             and not view_widget.render_widget.canvas.extent().isEmpty()]
+            
             if others_extents:
-                # set extent using the extent of the other valid view (or self) with at least one layer
-                extent = others_extents[0]
-                self.update_canvas_to(extent)
+                self.update_canvas_to(others_extents[0])
             elif self.canvas.extent().isEmpty():
-                # first layer to render
-                # set the extent using the extent of the Qgis project but first transform the crs if it is different
-                new_layer = valid_layers[0]
-                transform = QgsCoordinateTransform(new_layer.crs(), self.canvas.mapSettings().destinationCrs(), QgsProject.instance())
-                new_extent = transform.transformBoundingBox(new_layer.extent())
+                # first layer to render - set extent using the first valid layer
+                transform = QgsCoordinateTransform(
+                    valid_layers[0].crs(), 
+                    self.canvas.mapSettings().destinationCrs(), 
+                    QgsProject.instance()
+                )
+                new_extent = transform.transformBoundingBox(valid_layers[0].extent())
                 self.canvas.setExtent(new_extent)
 
             self.refresh()
