### 2025-10-28 135e8ad
Update tooltips for the registry items and tools related
