### 2025-10-29 7b816f2
Improve the handling of saving configuration with two buttons: one as the default for overwriting the current config file and another for selecting a new config file. Also, make improvements to the exit dialog with an option to save and exit.
