+++ gui/main_dialog.py
@@ -25,6 +25,7 @@ import tempfile
 from datetime import datetime
 from copy import deepcopy
 from pathlib import Path
+from html import escape
 import yaml
 try:
     from yaml import CSafeLoader as SafeLoader
@@ -34,7 +35,7 @@ except ImportError:
 from qgis.PyQt import uic
 from qgis.PyQt.QtCore import pyqtSignal, Qt, pyqtSlot, QTimer, QEvent
 from qgis.PyQt.QtWidgets import QMessageBox, QGridLayout, QFileDialog, QTableWidgetItem, QColorDialog, QWidget, QLabel, \
-    QComboBox, QDialog, QDockWidget, QFrame, QCheckBox
+    QComboBox, QDialog, QDockWidget, QFrame, QCheckBox, QApplication, QDialogButtonBox
 from qgis.core import Qgis, QgsMapLayerProxyModel, QgsRectangle, QgsPointXY, QgsCoordinateReferenceSystem, \
     QgsCoordinateTransform, QgsProject
 from qgis.PyQt.QtGui import QColor, QFont, QIcon
@@ -200,7 +201,9 @@ class ThRasEDialog(QDialog, FORM_CLASS):
         self.QPBtn_ApplyToEntireThematicRaster.clicked.connect(self.apply_to_entire_thematic_raster)
         self.apply_from_thematic_classes = ApplyFromThematicClasses()
         self.QPBtn_ApplyFromThematicClasses.clicked.connect(self.apply_from_thematic_classes_dialog)
-        self.SaveConfig.clicked.connect(self.file_dialog_save_thrase_config)
+        self.SaveConfig.clicked.connect(self.save_thrase_config)
+        self.SaveAsConfig.clicked.connect(self.file_dialog_save_thrase_config)
+        self.update_save_buttons_state()
 
         # ######### CCD plugin widget ######### #
         # set up the Continuous Change Detection widget
@@ -370,6 +373,7 @@ class ThRasEDialog(QDialog, FORM_CLASS):
                 self.QCBox_band_LayerToEdit.setCurrentIndex(yaml_config["thematic_file_to_edit"]["band"] - 1)
         # file config
         LayerToEdit.current.config_file = yaml_config["config_file"]
+        self.update_save_buttons_state()
         # symbology and pixel table
         if "symbology" in yaml_config:
             LayerToEdit.current.symbology = yaml_config["symbology"]
@@ -628,15 +632,45 @@ class ThRasEDialog(QDialog, FORM_CLASS):
             super(ThRasEDialog, self).keyPressEvent(event)
 
     def closeEvent(self, event):
-        # first prompt if dialog is opened
         if self.isVisible():
-            quit_msg = "Are you sure you want close the ThRasE plugin?"
-            reply = QMessageBox.question(None, 'Closing the ThRasE plugin',
-                                         quit_msg, QMessageBox.Yes, QMessageBox.No)
-            if reply == QMessageBox.No:
-                # don't close
+            msg_box = QMessageBox(self)
+            msg_box.setIcon(QMessageBox.Question)
+            msg_box.setWindowTitle(self.tr("Close ThRasE"))
+            msg_box.setTextFormat(Qt.RichText)
+            msg_box.setText("<p>{}</p>".format(
+                self.tr("Do you want to save the current configuration before exiting ThRasE?")
+            ))
+
+            layer_to_edit = LayerToEdit.current
+            if layer_to_edit and layer_to_edit.config_file:
+                config_path = layer_to_edit.config_file
+                try:
+                    if os.path.exists(config_path):
+                        saved_at = datetime.fromtimestamp(os.path.getmtime(config_path))
+                        config_path += " ({})".format(saved_at.strftime("%d %b %Y, %H:%M:%S"))
+                except OSError:
+                    pass
[…]
