### 2025-10-29 21bb246
If Thrase is open and Qgis is going to close, not block it and clossing together with Qgis
