### 2025-10-29 152f891
Added a tip note in the doc and other minos changes
