### 2025-10-31 0683e4a
Allow using the mouse pixel value tracking tool in the recode pixel table without needing to activate an editing tool
