### 2025-10-31 4a1681b
Adjustments in enable/disable some components when there is or not a later to edit already set
