### 2025-11-06 4dba70b
v25.11
