+++ metadata.txt
@@ -1,9 +1,9 @@
 [general]
 name=ThRasE
-description=A powerful QGIS plugin for fast thematic raster editing with recode pixel tables, modification tracking registry, and comprehensive navigation tools.
-about=ThRasE is a powerful and flexible QGIS Python plugin for fast thematic raster editing. It utilizes a recode pixel table to modify multiple classes simultaneously using various tools, including pixels, lines, polygons, and freehand drawing. It includes a registry system that tracks all modifications made during editing sessions. The plugin also features a navigation tool to facilitate systematic layer inspection.
+description=Open-source QGIS plugin for rigorous thematic raster editing with synchronized multi-panel views, recode pixel tables, modification tracking, tile navigation, and session persistence.
+about=ThRasE is an open-source QGIS plugin that provides rigorous editing for systematic post-classification correction and quality assurance of thematic rasters. It offers synchronized multi-panel views for side-by-side and overlap comparisons with reference data. It utilizes a recode pixel table to modify multiple classes simultaneously in one operation with various editing tool options. An integrated modification tracking system automatically records all changes, providing spatial feedback, transparency, and reproducibility. Systematic tile-based navigation ensures complete spatial coverage during inspection, while session persistence enables seamless multi-day workflows.
 
-version=25.8
+version=25.11
 qgisMinimumVersion=3.0
 qgisMaximumVersion=3.99
 
@@ -15,36 +15,39 @@ tracker=https://github.com/SMByC/ThRasE/issues
 repository=https://github.com/SMByC/ThRasE
 
 changelog=
-    25.8:
-    - Implement the registry tool widget for pixel modification tracking to review pixel changes over time
-    - Implement the button to export all registered modified pixels to a vector file, where each feature is a square polygon representing the pixel modified with the attributes: old_value, new_value, edit_date
-    - Implement saving and restoring of pixel logs and registry configuration and state from/to Thrase yaml file
-    - Add optional registry recording changes in the thematic class global editing tool
-    - Refactored the process for storing editing events and edited pixels using the PixelLog and EditLog classes, registering unique pixel edits by time and improving the overall editing workflow. Implemented QGIS’s geometry engine to optimize and improve performance of contains operations for polygon and freehand editing tools.
-    - Add auto-clear functionality for drawing tools
-    - Enhance layout or multiples view widgets with vertical and horizontal separators between them
-    - Refactor edit logging to use per-view history of edit actions for pixel, line, polygon, and freehand tools
-    - Add enable/disable functionality for the pixel change registry
-    - Only use GDAL for editing the whole-image function and catch any error in the process
-    - Set undo/redo buttons with a delay for pixel, line, polygon, and freehand to avoid issues in processing undo/redo actions for fast clicks
-    - Add the functionality to delete the entire pixel change registry
-    - Update navigation tiles color on color change without rebuild navigation
-    - Block wheel events on all QComboBoxes in Thrase dialog to prevent unwanted changes to selected items
-    - Fixing some issues in restoring configuration settings
-    - Disable and enable the registry button and widget based on layer selection
-    - Remove borders and adjust the layout margins of the CCD plugin for improvement integration
-    - Improve caching hash value in Pixel class
-    - Add group ID for pixel edits to track edit sessions
-    - Adjust edited pixels status messages
-    - Enable/disable editing toolbars if recode pixel table is empty
-    - Store/restore pixel tolerance handling and registry for current editing layer
-    - Set pixel tolerance for comparison based on pixel size
-    - Another rename related variables and methods
-    - Refactor typos in functions, variables, files names
-    - Avoid RuntimeError in edge cases with deleted plugin objects
-    - Improving and adjusting warning and error bar messages
-    - Minor UI and others
-    - Refine some icons
+    25.11:
[…]
