+++ _sources/example.md
@@ -1,23 +1,0 @@
-# Example
-
-```{image} images/tinigua.webp
-:width: 30%
-:align: center
-```
-
-This example provides a sample dataset to demonstrate the capabilities of ThRasE using a demo of a scenario. The data covers the Tinigua National Natural Park in Colombia, an area with a high rate of deforestation, and showcases forest land change between 2019 and 2020.
-
-The thematic map includes the following classes:
-*   Stable Forest (class 1)
-*   Non-Stable Forest (class 5)
-*   Deforestation (class 2)
-
-The example also includes mean composite maps for each year as a reference data.
-
-[Download the ThRasE Example ZIP file](https://github.com/SMByC/ThRasE/raw/refs/heads/main/docs/ThRasE%20Example.zip)
-
-Unzip the file and load the ThRasE session file (`.yaml`) in the init ThRasE dialog.
-
-```{important}
-All files in this example are for demonstration purposes only and do not represent official/real data.
-```
