+++ _sources/references.md
@@ -1,6 +1,0 @@
-# References
-
-```{bibliography}
-:style: unsrt
-```
-
