### 2025-11-28 5b28b1a
Update the PickerLineTool to work as a freehand drawing, the line will be drawn while the left mouse button is pressed, instead of by individual clicks
