### 2025-11-28 0eff489
Update the PickerPixelTool to support continuous editing while the left mouse button is pressed (like drawing)
