### 2025-11-30 a8a22f6
Update registry button tooltip when it is enabled/disabled
