+++ requirements_docs.txt
@@ -1,3 +1,3 @@
 # Documentation build dependencies for ThRasE
-jupyter-book
+jupyter-book<2
 jupytext
