+++ _sources/editing_tools.md
@@ -1,0 +1,18 @@
+# Editing and Reclassification Tools
+
+```{image} images/editing.webp
+:width: 90%
+:align: center
+```
+<br>
+ThRasE provides powerful editing capabilities for precise corrections of your thematic maps.
+<br>
+
+- You can **edit** at the pixel level or draw with lines, polygons, and freehand shapes.
+- The **recode pixel table** lets you define which classes should change to other classes to modify several classes at once in each operation.
+- You can start an edit from any **panel**, and the changes always apply to the thematic map, while keeping all your reference layers visible, preserving full visual context
+- Each tool keeps its own **undo** and **redo** history in each view, so you can edit safely and step back if something is not right.
+
+```{warning}
+After each editing operation, the layer is saved on the fly (overwritten) on disk. Make a backup copy before starting the editing process if you want to preserve the original layer.
+```
