+++ _sources/multi_view.md
@@ -1,0 +1,15 @@
+# Multi-View Configuration
+
+```{image} images/layer_toolbar.webp
+:width: 90%
+:align: center
+```
+<br>
+ThRasE lets you create multiple viewing panels arranged in a grid, all synchronized so they move together as you pan and zoom. Each panel can display up to three layers blended together with adjustable transparency.
+
+By arranging different layer combinations side by side and overlap layers with transparency, you can quickly spot problems like boundary mismatches, edge artifacts, and local inconsistencies that become obvious when you compare multiple views at once. This makes the review and correction process much faster because you can identify issues and fix them immediately without changing your view setup.
+<br>
+
+- **Grid configuration** can only be set when the plugin is first opened under the `New` tab
+- **Up to three layers per view** for comparing the thematic layer with reference data
+- **Adjustable opacity** for each layer according to your requirements
