+++ _sources/navigation.md
@@ -1,0 +1,12 @@
+# Systematic Inspection Navigator
+
+```{image} images/navigation.webp
+:width: 90%
+:align: center
+```
+<br>
+Systematic inspection is fundamental to quality assurance in thematic mapping. ThRasE provides a tile-based navigation tool that divides your area of interest into manageable units and guides you through them in a defined sequence to guarantee full spatial coverage. The navigation tool offers multiple configuration options, allowing you to adapt your review strategy to your data characteristics or geographic features. This structured approach transforms what could be an unstructured visual inspection into a methodical process, reducing the risk of missed areas and ensuring complete coverage during your editing and quality control work.
+
+```{tip}
+Choose a navigation pattern (horizontal or vertical) that aligns with the dominant features in your landscape for optimal workflow efficiency.
+```
