+++ _sources/ccd_integration.md
@@ -1,0 +1,20 @@
+# CCD Plugin Integration
+
+```{image} images/ccd_plugin.webp
+:width: 90%
+:align: center
+```
+<br>
+ThRasE integrates with the Continuous Change Detection (CCD) plugin to enhance your thematic raster editing workflow. The CCD plugin helps analyze trends and breakpoints of change over multi-year time series at specific locations through Google Earth Engine while you edit your thematic raster.
+<br>
+
+- **Temporal pattern analysis** at specific points of interest
+- **Multi-source support** for both Landsat and Sentinel data
+- **Spectral trend visualization** and detected breakpoints
+- **Change detection** - identification of gradual and abrupt changes over time
+
+Read more about the CCD plugin on its [official page](https://github.com/SMByC/CCD-Plugin)
+
+```{note}
+To use this feature, you must have the CCD Plugin installed in QGIS.
+```

[…]
