### 2025-12-01 1690c31
v25.12
