+++ docs/_toc.yml
@@ -18,7 +18,7 @@ parts:
   - caption: 'About'
     chapters:
       - file: about
-      - url: https://smbyc.github.io/ThRasE/paper.pdf
-        title: Paper pdf
+      # - url: https://smbyc.github.io/ThRasE/paper.pdf
+      #   title: Paper pdf
       - file: example
       - file: references
