### 2025-12-13 e34990b
Fix bibtext authors
