+++ paper/paper.bib
@@ -129,7 +129,7 @@
 }
 @misc{galindo:2014,
   title={P. Protocolo de procesamiento digital de imágenes para la cuantificación de la deforestación en Colombia. V 2.0},
-  author={Galindo G. and Espejo O. J. and Rubiano J. C. and Vergara L. K. and Cabrera E.},
+  author={Galindo, G. and Espejo, O. J. and Rubiano, J. C. and Vergara, L. K. and Cabrera, E.},
   year={2014},
   Publisher={Instituto de Hidrología, Meteorología y Estudios Ambientales – IDEAM. Bogotá D.C., Colombia.}
 }
