### 2025-12-14 e25f5b3
Minor fixes in tooltip/label typos
