### 2025-12-15 9cdfaeb
Use ltr and stable Qgis versions for tests
