### 2025-12-15 #9169 @jGaboardi
> Dear @jGaboardi, I want to make a minor update to the title to shorten it a bit:
>
> from: ThRasE: An Open-Source QGIS Plugin for Manual Post-Classification Correction and Quality Assurance of Thematic Maps
> to: ThRasE: A QGIS Plugin for Manual Post-Classification Correction and Quality Assurance of Thematic Maps
>
> Could you please confirm if we can do this without affecting the review process? Thanks

@XavierCLL – Yes, I think this is possible but I need to confirm with @kthyng.
