### 2025-12-15 cb6b27c
Minor title adjustment for the paper
