### 2025-12-15 #9634 @editorialbot
[REVIEW]: ThRasE: An Open-Source QGIS Plugin for Manual Post-Classification Correction and Quality Assurance of Thematic Maps
**Submitting author:** @XavierCLL (<a href="http://orcid.org/0000-0001-7593-6477">Xavier Corredor Llano</a>)
**Repository:** https://github.com/SMByC/ThRasE
**Branch with paper.md** (empty if default branch): main
**Version:** 25.8
