### 2025-12-16 #9634 @jGaboardi
**re: title**

@kthyng When I started this review the issue seems to have reverted to the original title[^1], not the updated one[^2]. Should I simply edit the title of this issue, or is there something more involved we need to do?

[^1]: ThRasE: An Open-Source QGIS Plugin for Manual Post-Classification Correction and Quality Assurance of Thematic Maps
[^2]: ThRasE: A QGIS Plugin for Manual Post-Classification Correction and Quality Assurance of Thematic Maps
