### 2025-12-24 7b7e5e5
Update Github workflow to trigger on changes in the docs directory
