+++ .github/workflows/build_docs.yml
@@ -4,6 +4,8 @@ on:
   push:
     branches:
       - main
+    paths:
+      - "docs/**"
 
 concurrency:
   group: "pages"
