### 2025-12-30 918cd4b
Refactor the way to get the unique pixel values from the raster in a memory efficient manner, to use it for the auto symbology function
