### 2025-12-31 fe4160c
Github actions and docker files updated, and clean up
