+++ .docker/run-docker-tests.sh
@@ -2,7 +2,7 @@
 
 set -e
 
-pushd /usr/src
-DEFAULT_PARAMS='./ThRasE/tests/ -v --qgis_disable_gui --qgis_disable_init'
-xvfb-run -a -s "-screen 0 1920x1080x24" pytest ${@:-`echo $DEFAULT_PARAMS`}
-popd
+cd /usr/src/ThRasE
+
+DEFAULT_PARAMS='./tests/ -v --qgis_disable_gui --qgis_disable_init'
+xvfb-run pytest ${@:-$DEFAULT_PARAMS}
