+++ requirements_docs.txt
@@ -1,3 +1,0 @@
-# Documentation build dependencies for ThRasE
-jupyter-book<2
-jupytext
