+++ .docker/docker-compose.gh.yml
@@ -4,7 +4,8 @@ services:
       context: ..
       dockerfile: ./.docker/Dockerfile
       args:
-        QGIS_TEST_VERSION: ${QGIS_TEST_VERSION}
+        QGIS_TEST_VERSION: ${QGIS_TEST_VERSION:-latest}
     tty: true
     volumes:
-      - ${GITHUB_WORKSPACE}:/usr/src/ThRasE
+      - ${GITHUB_WORKSPACE:-.}:/usr/src/ThRasE
+    working_dir: /usr/src/ThRasE
