### 2026-01-03 SMByC/ThRasE#7 @giswqs
JOSS Review Comments
Some comments for the JOSS paper: https://github.com/openjournals/joss-reviews/issues/9634

### Documentation

> Comment: The installation instructions can be found in the README, but not in the documentation: https://smbyc.github.io/ThRasE

> Currently, there is an [Example](https://smbyc.github.io/ThRasE/example.html) page in the About section, which is at the end of the documentation.  I would suggest adding an Installation section with example data at the beginning of the documentation instead of placing it at the very end.

> Missing

### Software paper

> Missing a **Summary** section at the beginning, although some of high-level functionality and purpose of the software can be found in the **Statement of need** section.

> Could not find information about the comparison to other packages.

> Some older references are missing DOIs.
