### 2026-01-04 104e2a7
Adjust metadata to CI
