### 2026-01-04 7ad405c
Added install instruction to the docs
