### 2026-01-04 2742c15
Update installation and example section in docs
