### 2026-01-05 6336fe1
Added summary for the paper and adjusted the short description
