### 2026-01-06 5daf854
Added a paragraph about the comparison with other packages in the statement of need
