+++ paper/paper.md
@@ -31,6 +31,8 @@ Thematic maps serve as indispensable analytical tools that simplify complex, con
 
 The value of a thematic map ultimately depends on its quality, which is primarily driven by the original data sources and classification methodologies. However, post-classification correction serves as a widely adopted final step to refine misclassifications and improve outcomes [@hutchinson:1982]. A common approach, particularly for remote sensing-based maps, involves integrating ancillary data and knowledge-based rules to resolve misclassifications, reduce commission errors, and enhance overall accuracy [@manandhar:2009; @thakkar:2017]. These corrections often rely on manual editing or semi-automated techniques guided by expert knowledge and reference data. This is precisely the context in which ThRasE was developed.
 
+Several tools offer raster editing and reclassification capabilities, including commercial packages such as ERDAS IMAGINE and ENVI with raster reclassification and pixel editing, and ArcGIS with its interactive pixel value manipulation tool. Open-source alternatives like GRASS GIS provide command-line reclassification, and the QGIS plugin Serval enables basic pixel-level raster modifications. While these tools support raster editing, none integrates synchronized multi-view comparison, systematic tile-based navigation, modification tracking, multiple editing tools, a recode table, and session persistence within a unified workflow.
+
 In Colombia, the Forest and Carbon Monitoring System (SMByC) of the Institute of Hydrology, Meteorology and Environmental Studies (IDEAM) is responsible for measuring and ensuring the accuracy of the official national forest figures. As part of the SMByC, we identified the need not only for review and editing capabilities but also for a dedicated tool to enable systematic and rigorous quality control of final products, thereby guaranteeing reliable results [@galindo:2014]. To address this gap, we developed ThRasE, an open-source QGIS plugin that provides flexible editing, visual inspection, and auditable quality assurance workflows.
 
 ThRasE offers a suite of editing tools, a modification tracking registry, systematic visual inspection features, and comparison capabilities with reference data. By incorporating ancillary data and knowledge-based rules, ThRasE enables expert-guided manual editing, facilitating targeted corrections of misclassifications and fine-tuning of class assignments. ThRasE has been used for multiple purposes worldwide, including manual post-processing to correct land-use/land-cover misclassifications via photo-interpretation of satellite image mosaics [@vallet:2024; @rayner:2022]; knowledge-based manual correction of misclassifications and residual errors [@rayner:2021effects; @senterre:2023; @bladh:2024]; manual classification/reclassification of pixels for land-cover reconstruction to reconcile multi-date imagery [@gunawan:2023]; annotation of historical aerial imagery for model calibration and fine-tuning [@eyster:2024]; as a validation/visual QA step [@hariyanto:2024]; refinement of agricultural maps [@gandharum:2025; @rayner:2021historical]; among other applications [@hariyanto:2022; @dupuy:2024; @queiroga:2020; @nunes:2024]. Furthermore, ThRasE is an official component of our Digital Image Processing Protocol for Quantifying Deforestation in Colombia [@galindo:2014].
