### 2026-01-06 a270777
Updated bibliography entries
