+++ paper/paper.bib
@@ -43,15 +43,6 @@
   publisher={Elsevier},
   doi={10.1016/j.ejrs.2016.11.006}
 }
-@article{hutchinson:1982,
-  title={Classification improvement},
-  author={Hutchinson, Charles F},
-  journal={Photogrammetric Engineering and Remote Sensing},
-  volume={44},
-  number={1},
-  pages={123--130},
-  year={1982}
-}
 @article{foody:2001,
   title={GIS: the accuracy of spatial data revisited},
   author={Foody, Giles M},
@@ -71,6 +62,7 @@
   number={10},
   pages={1423--1430},
   year={1987}
+  url={https://www.academia.edu/download/90543545/1987_oct_1423-1430.pdf}
 }
 @book{maling:2013,
   title={Measurements from maps: principles and methods of cartometry},
@@ -126,12 +118,14 @@
   number={02},
   pages={259--263},
   year={2012}
+  url={https://njcmindia.com/index.php/file/article/view/1692}
 }
 @misc{galindo:2014,
-  title={P. Protocolo de procesamiento digital de imágenes para la cuantificación de la deforestación en Colombia. V 2.0},
+  title={Protocolo de procesamiento digital de imágenes para la cuantificación de la deforestación en Colombia. V 2.0},
   author={Galindo, G. and Espejo, O. J. and Rubiano, J. C. and Vergara, L. K. and Cabrera, E.},
   year={2014},
-  Publisher={Instituto de Hidrología, Meteorología y Estudios Ambientales – IDEAM. Bogotá D.C., Colombia.}
+  Publisher={Instituto de Hidrología, Meteorología y Estudios Ambientales – IDEAM. Bogotá D.C., Colombia.},
+  url={https://www.ideam.gov.co/sites/default/files/archivos/protocolo_de_pdi_para_la_cuantificacion_de_la_deforestacion_en_colombia_v2_1.pdf}
 }
 @article{eyster:2024,
   title={Space-for-time substitutions exaggerate urban bird--habitat ecological relationships},
@@ -165,12 +159,13 @@
   publisher={MDPI},
   doi={10.3390/su15097722}
 }
-@article{senterre:2023,
+@report{senterre:2023,
   title={Flora and vegetation of Assomption Island (Seychelles)},
   author={Senterre, Bruno and Mederic, Emma and Berke, Gregory},
   journal={Island Conservation Society (ICS), Pointe Larue, PO Box},
   volume={775},
-  year={2023}
+  year={2023},
+  url={https://www.researchgate.net/profile/Bruno-Senterre/publication/378943939_Flora_and_vegetation_of_Assomption_Island_Seychelles/links/65f2f44fc05fd2688010af87/Flora-and-vegetation-of-Assomption-Island-Seychelles.pdf}
 }
 @article{hariyanto:2024,
   title={Comparison of Land Cover Change Prediction Models: A Case Study in Kedungkandang District, Malang City},
@@ -205,14 +200,16 @@
   title={Urban Tree Mapping using airborne LiDAR: Analysing vegetation changes between 2010, 2017, and 2022 A Gothenburg case study},
   author={Bladh, Evelina},
   year={2024},
-  school={University of Gothenburg}
+  school={University of Gothenburg},
+  url={https://hdl.handle.net/2077/81972}
 }
 @phdthesis{rayner:2022,
   author={Rayner, Maxwell John},
-  title={{Synergies and trade-offs in ecosystem services across the Upper Welland catchment}},
[…]
