+++ paper/paper.md
@@ -29,7 +29,7 @@ ThRasE is an open-source QGIS plugin designed for systematic review, manual corr
 
 Thematic maps serve as indispensable analytical tools that simplify complex, continuous spatial data into visually accessible representations for a wide range of disciplines and applications. However, mapping workflows inevitably introduce errors, making quality assurance essential for reliable spatial data and informed decision-making [@foody:2001]. Because maps are models of reality, they inevitably contain errors introduced by projection, compilation, and symbolization [@lightfoot:1987; @maling:2013]. Many issues originate in the production process itself. For example, satellite-based classifications (e.g., supervised or unsupervised workflows) can introduce errors of omission or commission, scale mismatches, or temporal misalignments with ground conditions [@Schowengerdt:2007]. These limitations do not undermine the value of thematic maps; rather, they highlight the importance of rigorous quality assurance. Best practices include conducting formal accuracy assessments [@congalton:2019], transparently reporting error and uncertainty [@maceachren:1992], and, where appropriate, applying targeted post-classification corrections to address misclassifications [@manandhar:2009].
 
-The value of a thematic map ultimately depends on its quality, which is primarily driven by the original data sources and classification methodologies. However, post-classification correction serves as a widely adopted final step to refine misclassifications and improve outcomes [@hutchinson:1982]. A common approach, particularly for remote sensing-based maps, involves integrating ancillary data and knowledge-based rules to resolve misclassifications, reduce commission errors, and enhance overall accuracy [@manandhar:2009; @thakkar:2017]. These corrections often rely on manual editing or semi-automated techniques guided by expert knowledge and reference data. This is precisely the context in which ThRasE was developed.
+The value of a thematic map ultimately depends on its quality, which is primarily driven by the original data sources and classification methodologies. However, post-classification correction serves as a widely adopted final step to refine misclassifications and improve outcomes. A common approach, particularly for remote sensing-based maps, involves integrating ancillary data and knowledge-based rules to resolve misclassifications, reduce commission errors, and enhance overall accuracy [@manandhar:2009; @thakkar:2017]. These corrections often rely on manual editing or semi-automated techniques guided by expert knowledge and reference data. This is precisely the context in which ThRasE was developed.
 
 Several tools offer raster editing and reclassification capabilities, including commercial packages such as ERDAS IMAGINE and ENVI with raster reclassification and pixel editing, and ArcGIS with its interactive pixel value manipulation tool. Open-source alternatives like GRASS GIS provide command-line reclassification, and the QGIS plugin Serval enables basic pixel-level raster modifications. While these tools support raster editing, none integrates synchronized multi-view comparison, systematic tile-based navigation, modification tracking, multiple editing tools, a recode table, and session persistence within a unified workflow.
 
