+++ paper/paper.bib
@@ -61,7 +61,7 @@
   volume={53},
   number={10},
   pages={1423--1430},
-  year={1987}
+  year={1987},
   url={https://www.academia.edu/download/90543545/1987_oct_1423-1430.pdf}
 }
 @book{maling:2013,
@@ -117,7 +117,7 @@
   volume={3},
   number={02},
   pages={259--263},
-  year={2012}
+  year={2012},
   url={https://njcmindia.com/index.php/file/article/view/1692}
 }
 @misc{galindo:2014,
