### 2026-01-07 #9634 @XavierCLL
> [@jGaboardi](https://github.com/jGaboardi) I have completed my review. Thank you to [@XavierCLL](https://github.com/XavierCLL) addressing my comments at [SMByC/ThRasE#7](https://github.com/SMByC/ThRasE/issues/7)

@giswqs thanks so much for your time and for your feedback on this review :)
