### 2026-01-09 28d5835
Fix bibtex entry
