+++ paper/paper.bib
@@ -159,10 +159,10 @@
   publisher={MDPI},
   doi={10.3390/su15097722}
 }
-@report{senterre:2023,
+@techreport{senterre:2023,
   title={Flora and vegetation of Assomption Island (Seychelles)},
   author={Senterre, Bruno and Mederic, Emma and Berke, Gregory},
-  journal={Island Conservation Society (ICS), Pointe Larue, PO Box},
+  institution={Island Conservation Society (ICS), Pointe Larue, PO Box},
   volume={775},
   year={2023},
   url={https://www.researchgate.net/profile/Bruno-Senterre/publication/378943939_Flora_and_vegetation_of_Assomption_Island_Seychelles/links/65f2f44fc05fd2688010af87/Flora-and-vegetation-of-Assomption-Island-Seychelles.pdf}
