### 2026-01-19 d726987
Fix releasing with creating GitHub release in workflow
