+++ .github/workflows/release.yml
@@ -30,6 +30,11 @@ jobs:
       - name: Install qgis-plugin-ci
         run: pip install qgis-plugin-ci
 
+      - name: Create GitHub Release
+        run: gh release create ${{ github.ref_name }} --title "${{ github.ref_name }}" --notes "Automated release"
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
       - name: Package and release
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
