### 2026-02-01 #9634 @Narayana-Rao
@XavierCLL, the plugin offers extensive functionality and is well built. The paper itself is also well written. However, the documentation is limited. Specifically, the [example usage](https://smbyc.github.io/ThRasE/install_and_example.html#example) does not provide sufficient detail. I would expect to see step‑by‑step instructions for a specific task, which would be especially useful for someone taking a first look at the plugin.

In my opinion, documentation is equally important as the software itself, since the primary goal is to make it accessible and usable by others.

During testing, QGIS crashed once while integrating the [CCD‑plugin](https://plugins.qgis.org/plugins/CCD_Plugin/) for the first time, but worked fine afterwards.
Environment: QGIS 3.44.7 on Ubuntu 24.04.3 LTS.
