### 2026-02-03 e915495
Enhance example documentation with detailed steps
