+++ _sources/about.md
@@ -6,7 +6,5 @@ We identified the need for a dedicated tool that could support systematic and ri
 
 ThRasE is an official component of our Digital Image Processing Protocol for Quantifying Deforestation in Colombia {cite}`galindo:2014`, where it plays a central role in our quality assurance workflows. We use ThRasE to conduct systematic reviews of land cover classifications, correct misclassifications through expert visual interpretation of satellite imagery, and ensure our forest monitoring products meet the accuracy standards required for national reporting.
 
-ThRasE has been adopted by research teams, monitoring programs, and mapping projects worldwide for diverse applications including manual post-processing to correct land-use/land-cover misclassifications via photo-interpretation of satellite image mosaics {cite}`vallet:2024,rayner:2022`; knowledge-based manual correction of misclassifications and residual errors {cite}`rayner:2021effects,senterre:2023,bladh:2024`; manual classification/reclassification of pixels for land-cover reconstruction to reconcile multi-date imagery {cite}`gunawan:2023`; annotation of historical aerial imagery for model calibration and fine-tuning {cite}`eyster:2024`; as a validation/visual QA step {cite}`hariyanto:2024`; refinement of agricultural maps {cite}`gandharum:2025,rayner:2021historical`; among other applications {cite}`hariyanto:2022,dupuy:2024,queiroga:2020,nunes:2024`.
-
 Author and developer: *Xavier C. Llano* *<xavier.corredor.llano@gmail.com>*  
 Theoretical support, testing, and product verification: SMByC-PDI group
