### 2026-02-03 24fc651
Minor types and example update
