### 2026-02-03 #9634 @XavierCLL
Dear @Narayana-Rao thanks so much for your time and for your feedback.

You’re right, the demo page is very basic and doesn’t show the main features of the plugin. I’ve updated the documentation with a step-by-step guide to walk users through the plugin’s functionality:

https://smbyc.github.io/ThRasE/install_and_example.html

Please check it out and let me know your thoughts. Thank you!
