### 2026-02-16 #9634 @jGaboardi
@XavierCLL – please review the *Additional Author Tasks After Review is Complete* [above](https://github.com/openjournals/joss-reviews/issues/9634#:~:text=Additional%20Author%20Tasks%20After%20Review%20is%20Complete) for those 5 bullet points.

Once you have completed those bullets, ping me here and I will start on the *Editor Tasks Prior to Acceptance*.
