### 2026-02-17 #9634 @XavierCLL
Dear @jGaboardi

I have checked all the items in the Post-Review Checklist, however I cannot set neither the DOI nor the version:

DOI: 10.5281/zenodo.18671535
Version: 26.2

Thanks
