### 2026-02-17 #9634 @XavierCLL
> [@XavierCLL](https://github.com/XavierCLL) – Please update your Zenodo Author entry to have your associated ORCiD. Here is an example from previously published archive - https://zenodo.org/records/17211085

Done
