### 2026-02-17 #9634 @jGaboardi
Nice work on this @XavierCLL - now this is over to the track editor for further processing.

And thank you @giswqs & @Narayana-Rao for your timely and thoughtful reviews!
