### 2026-02-20 #9634 @kthyng
@XavierCLL Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "gothenberg" but please check for more.
