### 2026-02-20 #9634 @jGaboardi
> [@XavierCLL](https://github.com/XavierCLL) Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "gothenberg" but please check for more.

ah, my mistake for missing this @kthyng
