### 2026-02-20 f3b2faf
Fix capitalization in bibliography
