+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @incollection{Schowengerdt:2007,
-    title = {CHAPTER 9 - Thematic Classification},
+    title = {{CHAPTER} 9 - Thematic Classification},
     editor = {Robert A. Schowengerdt},
     booktitle = {Remote Sensing (Third Edition)},
     publisher = {Academic Press},
@@ -15,14 +15,14 @@
 @article{maceachren:1992,
   title={Visualizing uncertain information},
   author={MacEachren, Alan M},
-  journal={Cartographic perspectives},
+  journal={Cartographic Perspectives},
   number={13},
   pages={10--19},
   year={1992},
   doi={10.14714/CP13.1000}
 }
 @article{manandhar:2009,
-  title={Improving the accuracy of land use and land cover classification of Landsat data using post-classification enhancement},
+  title={Improving the accuracy of land use and land cover classification of {Landsat} data using post-classification enhancement},
   author={Manandhar, Ramita and Odeh, Inakwu OA and Ancev, Tiho},
   journal={Remote Sensing},
   volume={1},
@@ -33,7 +33,7 @@
   doi={10.3390/rs1030330}
 }
 @article{thakkar:2017,
-  title={Post-classification corrections in improving the classification of Land Use/Land Cover of arid region using RS and GIS: The case of Arjuni watershed, Gujarat, India},
+  title={Post-classification corrections in improving the classification of Land Use/Land Cover of arid region using {RS} and {GIS}: The case of {Arjuni} watershed, {Gujarat}, {India}},
   author={Thakkar, Amee K and Desai, Venkappayya R and Patel, Ajay and Potdar, Madhukar B},
   journal={The Egyptian Journal of Remote Sensing and Space Science},
   volume={20},
@@ -44,7 +44,7 @@
   doi={10.1016/j.ejrs.2016.11.006}
 }
 @article{foody:2001,
-  title={GIS: the accuracy of spatial data revisited},
+  title={{GIS}: the accuracy of spatial data revisited},
   author={Foody, Giles M},
   journal={Progress in Physical Geography},
   volume={25},
@@ -75,11 +75,11 @@
   title={Assessing the accuracy of remotely sensed data: principles and practices},
   author={Congalton, Russell G and Green, Kass},
   year={2019},
-  publisher={CRC press},
+  publisher={CRC Press},
   doi={10.1201/9781420048568}
 }
 @article{vallet:2024,
-  title={Generating high-resolution land use and land cover maps for the greater Mari{\~n}o watershed in 2019 with machine learning},
+  title={Generating high-resolution land use and land cover maps for the greater {Mari{\~n}o} watershed in 2019 with machine learning},
   author={Vallet, Am{\'e}line and Dupuy, St{\'e}phane and Verlynde, Matthieu and Gaetano, Raffaele},
   journal={Scientific Data},
   volume={11},
@@ -90,7 +90,7 @@
   doi={10.1038/s41597-024-03750-x}
 }
 @article{biljecki:2021,
-  title={Street view imagery in urban analytics and GIS: A review},
+  title={Street view imagery in urban analytics and {GIS}: A review},
   author={Biljecki, Filip and Ito, Koichi},
   journal={Landscape and Urban Planning},
   volume={215},
@@ -102,7 +102,7 @@
 @article{sishodia:2020,
   title={Applications of remote sensing in precision agriculture: A review},
   author={Sishodia, Rajendra P and Ray, Ram L and Singh, Sudhir K},
[…]
