### 2026-02-20 #9634 @XavierCLL
@kthyng thanks for your review, I fixed capitalization in [this commit](https://github.com/SMByC/ThRasE/commit/f3b2fafc517902003881ee4d20d87fc4fc7a337f), could you please check if everything is okay now?
