### 2026-02-20 #9634 @jGaboardi
@XavierCLL – Looks like some are still not properly cased. For example [https://doi.org/10.14710/geoplanning](https://ejournal.undip.ac.id/index.php/geoplanning/article/view/47318).

```
Hariyanto, A. D., Yudono, A., & Wicaksono, A. D. (2024). Comparison of land cover
    change prediction models: A case study in Kedungkandang District, Malang City.
    Geoplanning: Journal of Geomatics and Planning, 11(1), 85–98.
    https://doi.org/10.14710/geoplanning. 11.1.85-98
```

Best to check all the references to ensure proper casing.
