### 2026-02-21 4e217f5
Fix casing in bibliography
