+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @incollection{Schowengerdt:2007,
-    title = {{CHAPTER} 9 - Thematic Classification},
+    title = {{CHAPTER} 9 - {Thematic Classification}},
     editor = {Robert A. Schowengerdt},
     booktitle = {Remote Sensing (Third Edition)},
     publisher = {Academic Press},
@@ -121,7 +121,7 @@
   url={https://njcmindia.com/index.php/file/article/view/1692}
 }
 @misc{galindo:2014,
-  title={Protocolo de procesamiento digital de imágenes para la cuantificación de la deforestación en {Colombia}. {V} 2.0},
+  title={{Protocolo de procesamiento digital de imágenes para la cuantificación de la deforestación en Colombia. V 2.0}},
   author={Galindo, G. and Espejo, O. J. and Rubiano, J. C. and Vergara, L. K. and Cabrera, E.},
   year={2014},
   publisher={Instituto de Hidrología, Meteorología y Estudios Ambientales – IDEAM. Bogotá D.C., Colombia.},
@@ -168,7 +168,7 @@
   url={https://www.researchgate.net/profile/Bruno-Senterre/publication/378943939_Flora_and_vegetation_of_Assomption_Island_Seychelles/links/65f2f44fc05fd2688010af87/Flora-and-vegetation-of-Assomption-Island-Seychelles.pdf}
 }
 @article{hariyanto:2024,
-  title={Comparison of Land Cover Change Prediction Models: A Case Study in {Kedungkandang District}, {Malang City}},
+  title={{Comparison of Land Cover Change Prediction Models: A Case Study in Kedungkandang District, Malang City}},
   author={Hariyanto, Annisa Dira and Yudono, Adipandang and Wicaksono, Agus Dwi},
   journal={Geoplanning: Journal of Geomatics and Planning},
   volume={11},
@@ -179,7 +179,7 @@
   doi={10.14710/geoplanning.11.1.85-98}
 }
 @article{gandharum:2025,
-  title={Advancing Land Use Modeling with Rice Cropping Intensity: A Geospatial Study on the Shrinking Paddy Fields in {Indonesia}},
+  title={{Advancing Land Use Modeling with Rice Cropping Intensity: A Geospatial Study on the Shrinking Paddy Fields in Indonesia}},
   author={Gandharum, Laju and Hartono, Djoko Mulyo and Sadmono, Heri and Sanjaya, Hartanto and Sumargana, Lena and Kusumawardhani, Anindita Diah and Alhasanah, Fauziah and Sencaki, Dionysius Bryan and Setyaningrum, Nugraheni},
   journal={Geographies},
   volume={5},
@@ -190,14 +190,14 @@
   doi={10.3390/geographies5030031}
 }
 @article{rayner:2021historical,
-  title={Historical Declines in Semi-Natural Habitat Across an Agricultural Landscape Drive Modelled Losses in Wild Bee Abundance},
+  title={{Historical Declines in Semi-Natural Habitat Across an Agricultural Landscape Drive Modelled Losses in Wild Bee Abundance}},
   author={Rayner, Max and Whelan, Mick and Stoate, Chris and Szczur, John and Balzter, Heiko and Jones, Laurence},
   year={2021},
   journal={SSRN},
   doi={10.2139/ssrn.3969794}
 }
 @mastersthesis{bladh:2024,
-  title={Urban Tree Mapping using airborne {LiDAR}: Analysing vegetation changes between 2010, 2017, and 2022 A {Gothenburg} case study},
[…]
