### 2026-02-21 29c605d
Fix title-case bibliography v2
