+++ paper/paper.bib
@@ -13,7 +13,7 @@
     author = {Robert A. Schowengerdt},
 }
 @article{maceachren:1992,
-  title={Visualizing uncertain information},
+  title={{Visualizing Uncertain Information}},
   author={MacEachren, Alan M},
   journal={Cartographic Perspectives},
   number={13},
@@ -22,7 +22,7 @@
   doi={10.14714/CP13.1000}
 }
 @article{manandhar:2009,
-  title={Improving the accuracy of land use and land cover classification of {Landsat} data using post-classification enhancement},
+  title={{Improving the Accuracy of Land Use and Land Cover Classification of Landsat Data Using Post-Classification Enhancement}},
   author={Manandhar, Ramita and Odeh, Inakwu OA and Ancev, Tiho},
   journal={Remote Sensing},
   volume={1},
@@ -33,7 +33,7 @@
   doi={10.3390/rs1030330}
 }
 @article{thakkar:2017,
-  title={Post-classification corrections in improving the classification of Land Use/Land Cover of arid region using {RS} and {GIS}: The case of {Arjuni} watershed, {Gujarat}, {India}},
+  title={Post-classification corrections in improving the classification of {Land Use/Land Cover} of arid region using {RS} and {GIS}: The case of {Arjuni} watershed, {Gujarat}, {India}},
   author={Thakkar, Amee K and Desai, Venkappayya R and Patel, Ajay and Potdar, Madhukar B},
   journal={The Egyptian Journal of Remote Sensing and Space Science},
   volume={20},
@@ -72,7 +72,7 @@
   doi={10.1016/0160-9327(89)90096-3}
 }
 @book{congalton:2019,
-  title={Assessing the accuracy of remotely sensed data: principles and practices},
+  title={{Assessing the Accuracy of Remotely Sensed Data: Principles and Practices}},
   author={Congalton, Russell G and Green, Kass},
   year={2019},
   publisher={CRC Press},
@@ -100,7 +100,7 @@
   doi={10.1016/j.landurbplan.2021.104217}
 }
 @article{sishodia:2020,
-  title={Applications of remote sensing in precision agriculture: A review},
+  title={{Applications of Remote Sensing in Precision Agriculture: A Review}},
   author={Sishodia, Rajendra P and Ray, Ram L and Singh, Sudhir K},
   journal={Remote Sensing},
   volume={12},
@@ -149,7 +149,7 @@
   doi={10.1016/j.ecolind.2021.108463}
 }
 @article{gunawan:2023,
-  title={From laborers to coffee farmers: collaborative forest management in {West Java}, {Indonesia}},
+  title={{From Laborers to Coffee Farmers: Collaborative Forest Management in West Java, Indonesia}},
   author={Gunawan, Budhi and Abdoellah, Oekan Soekotjo and Hadi, Firman and Alifi, Gianrico Juan and Suhendi, Riky Novalia and Aisharya, Inas Yaumi and Gunawan, Wahyu},
   journal={Sustainability},
   volume={15},
@@ -197,7 +197,7 @@
   doi={10.2139/ssrn.3969794}
 }
 @mastersthesis{bladh:2024,
-  title={Urban {Tree Mapping} using airborne {LiDAR}: Analysing vegetation changes between 2010, 2017, and 2022 A {Gothenburg} case study},
+  title={Urban {Tree Mapping} using airborne {LiDAR}: Analysing vegetation changes between 2010, 2017, and 2022 {A} {Gothenburg} case study},
   author={Bladh, Evelina},
   year={2024},
   school={University of Gothenburg},
