### 2026-02-21 e10f099
Fix title-case bibliography v3
