+++ paper/paper.bib
@@ -228,7 +228,7 @@
   url={https://agritrop.cirad.fr/608783/}
 }
 @mastersthesis{queiroga:2020,
-  title={{Interfer{\^e}ncia das autoestradas na conectividade funcional da paisagem para os anf{\'\i}bios do Peloponeso na Gr{\'e}cia}},
+  title={{Interfer{\^e}ncia das Autoestradas na Conectividade Funcional da Paisagem para os Anf{\'\i}bios do Peloponeso na Gr{\'e}cia}},
   author={Queiroga, Fabr{\'\i}cio Mendes},
   year={2020},
   school={Universidade do Porto (Portugal)},
@@ -240,7 +240,7 @@ author = {Nunes Machado, Alan and Costa, Kylner and Wachholz, Flávio},
 year = {2024},
 month = {02},
 pages = {103-112},
-title = {{Sensoriamento remoto aplicado à prática de campo no ensino de hidrografia e geomorfologia fluvial: Uma análise do arquipélago de Anavilhanas (Amazônas-Brasil)}},
+title = {{SENSORIAMENTO REMOTO APLICADO À PRÁTICA DE CAMPO NOENSINO DE HIDROGRAFIA E GEOMORFOLOGIA FLUVIAL:UMA ANÁLISE DO ARQUIPÉLAGO DE ANAVILHANAS (AMAZONAS-BRASIL)}},
 booktitle = {Remote Sensing and Field Practice in Geography and Geomorphology},
 publisher = {Editora Científica},
 isbn = {978-85-5467-374-1},
