### 2026-02-21 #9634 @XavierCLL
Dear @jGaboardi you are right there were some casing issues in some titles, I checked one by one, I think now it is fixed, thanks
