### 2026-02-22 #9634 @XavierCLL
Wow, @jGaboardi! what a good eye! Thanks for the PR!
