### 2026-03-06 2adc503
Improve and fix the way to chek the data type of thematic layer for different Qgis versions and also support Int8 as valid integer data type
